Improve error message for attempted directory unlinks (Patch from Torsten Rehn) (Resolves Ticket #657)
git-svn-id: https://svn.mcs.anl.gov/repos/bcfg/trunk/bcfg2@5277 ce84e21b-d406-0410-9b95-82705330c041

+++ src/lib/Client/Tools/POSIX.py
@@ -279,6 +279,9 @@ class POSIX(Bcfg2.Client.Tools.Tool):
         if entry.get('prune', 'false') == 'true' and entry.get("qtest"):
             for pname in entry.get("qtest").split(":"):
                 ulfailed = False
+                if os.path.isdir(pname): 
+                    self.logger.info("Not removing extra directory %s, please check and remove manually" % pname)
+                    continue
                 try:
                     self.logger.debug("Unlinking file %s" % pname)
                     os.unlink(pname)
